+++ docker/build.d/shop/800_finishing_tasks.sh
@@ -3,7 +3,7 @@
 # fix error with missing event log dir
 # TODO: configure log destination to /data/logs/
 sectionText "Creating required directories"
-mkdir -vp $WORKDIR/data/$SPRYKER_SHOP_CC/logs/
+mkdir -vp $WORKDIR/data/$SPRYKER_SHOP_CC/logs/ZED
 mkdir -vp /data/logs
 
 # TODO: increase security by making this more granular