+++ docker/common.inc.sh
@@ -73,6 +73,27 @@ writeErrorMessage() {
   fi
 }
 
+#
+#  INIT helper functions
+#
+configure_jenkins() {
+  sectionText "Configuring jenkins as the cronjob handler"
+  
+  wait_for_http_service http://$JENKINS_HOST:$JENKINS_PORT
+
+  # FIXME [bug01] until the code of the following cronsole command completely
+  # relies on API calls, we need to workaround the issue with missing local
+  # jenkins job definitions.
+  mkdir -p /tmp/jenkins/jobs
+  # Generate Jenkins jobs configuration
+  $CONSOLE setup:jenkins:generate
+}
+
+configure_crond() {
+  sectionText "Configuring crond as the cronjob handler"
+  php $WORKDIR/docker/contrib/gen_crontab.php
+}
+
 
 install_packages() {
   local INSTALL_FLAGS="--no-cache"
@@ -153,7 +174,7 @@ exec_scripts() {
   if [ -d "$directory" ]; then
     
     # provide script counting to inform the user about how many steps are available
-    local available_scripts=`find $directory -type f -name '*.sh' | sort`
+    local available_scripts=`find $directory -type f -name '*.sh' -or -name '*.php' | sort`
     local scripts_count=`echo "$available_scripts" | wc -l`
     local scripts_counter=1